requirements.txt
appnope==0.1.0
click==6.7
decorator==4.0.11
Flask==0.12.2
GroupyAPI==0.7.1
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.0
pandas==0.20.2
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==2.5.3
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.3.0
simplegeneric==0.8.1
six==1.10.0
slackclient==1.0.6
traitlets==4.3.2
wcwidth==0.1.7
websocket-client==0.40.0
Werkzeug==0.12.2
